reeniciando o projeto

+++ CadastroDeClientes/CadastroDeClientes/src/main/java/com/allanisaac_dev/CadastroDeClientes/Cliente/ClienteService.java
@@ -1,47 +1,58 @@
 package com.allanisaac_dev.CadastroDeClientes.Cliente;
 
-
-import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.PathVariable;
 
-import javax.swing.text.html.Option;
 import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 @Service
-@RequiredArgsConstructor
-
 public class ClienteService {
 
-    private final ClienteRepository clienteRepository;
-    private final ClienteMapper clienteMapper;
 
-    //Lista todos os clientes cadastrados
-    public List<ClienteDTO> listarTodosClientes(){
-        List<ClienteModel> clienteModel =clienteRepository.findAll();
-        return clienteModel.stream()
-                .map(clienteMapper::map).collect(Collectors.toList());
+    //Essa anotação faz injeção de dependencia, o que faz com que a gente não precise escrever construtor
+    @Autowired
+    private ClienteRepository clienteRepository;
+
+
+    //Cadastra cliente
+    public ClienteModel cadastrarCliente(ClienteModel clienteModel){
+        return clienteRepository.save(clienteModel);
     }
 
-    //Faz uma pesquisa por nome
-    public List<ClienteDTO> listaClientePorNome(String nome){
-        List<ClienteModel> clientePorNome = clienteRepository.findByNome(nome);
-        return clientePorNome.stream()
-                .map(clienteMapper::map).collect(Collectors.toList());
+
+    //Lista todos os clientes
+    public List<ClienteModel> listarClientes(){
+        return clienteRepository.findAll();
     }
 
-    //Cadastra um novo cliente
-    public ClienteDTO cadastrarCliente(ClienteDTO clienteDTO){
-        ClienteModel clienteModel = clienteMapper.map(clienteDTO);
-        clienteModel = clienteRepository.save(clienteModel);
-        return clienteMapper.map(clienteModel);
+    //Lista cliente por ID
+    public ClienteModel listarClientePorId(Long id){
+        Optional<ClienteModel> cliente = clienteRepository.findById(id);
+        return cliente.orElse(null);
     }
 
-    //Deletar cliente
-    public void deletarClientePorID(Long id){
-        clienteRepository.deleteById(id);
+    //Editar/Atualiza cliente por ID
+    public ClienteModel atualizarCliente(Long id, ClienteModel clienteAtualizado){
+        if (clienteRepository.existsById(id)){
+            clienteAtualizado.setId(id);
+            return clienteRepository.save(clienteAtualizado);
+        }
+        return null;
+
     }
+
+    //Deleta cliente por ID
+    public void deletarCliente(Long id){
+        Optional<ClienteModel> cliente = clienteRepository.findById(id);
+
+        if (cliente.isPresent()){
+            clienteRepository.deleteById(id);
+        }else {
+            System.out.println("Cliente não encontrado" );
+        }
+
+    }
+
 
 }
